Split ScrapperController into TgChatController and LinksController
